requirements.txt
gasgiref==3.8.1
Django==5.2.2
django-modeltranslation==0.19.15
pillow==11.2.1
sqlparse==0.5.3
